// src/ledger/ledger.enums.ts
export enum LedgerAccountType {
  ASSET = 'ASSET',
  LIABILITY = 'LIABILITY',
  EQUITY = 'EQUITY',
  REVENUE = 'REVENUE',
  EXPENSE = 'EXPENSE',
}

export enum LedgerNormalBalance {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}

export enum LedgerEntryDirection {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}

export enum LedgerJournalStatus {
  POSTED = 'POSTED',
}